requirements.txt
aiogram==3.1.1
python-dotenv
pydantic==2.3
redis==5.0.8